게시판 댓글, 멤버 이름 연결 (#20)
* feat: post에 memberId 연결

* design: 게시글 리스트, 게시글 상세에 작성자 표시되도록 변경

* chore: 작성자를 회원이름으로 표시되도록 변경

* front 작업 및 검증 과정 (#18)

* feat: 로그인,등록 성공 alert 추가

* feat: alert창 수정

* feat: member check 로직 추가

* feat: register front 검증 과정

* feat: memberId로 member 찾기 추가

* feat: 게시글에 댓글 추가

* feat: 게시판에도 회원 이름 뜨도록 추가

* refactor: 게시글 소유자 확인 로직을 서비스 계층으로 이동

* refactor: 의존성 일관성 유지를 위해 생성자 주입 방식으로 변경

* refactor: 메소드의 역할을 정확하게 하기 위해 로직 분리

* feat: 게시글 수정, 삭제 기능 구현

* feat: 작성자만 수정, 삭제 가능하도록 구현

* design: 게시판, 게시글 디자인 수정

* design: 게시글 디자인 수정

* design: 게시글 디자인 수정

---------

+++ src/main/java/com/besp/likebesp1/post/controller/PostController.java
@@ -2,29 +2,67 @@ package com.besp.likebesp1.post.controller;
 
 import com.besp.likebesp1.board.entity.BoardDto;
 import com.besp.likebesp1.board.service.BoardService;
+import com.besp.likebesp1.cmnt.dto.CmntDto;
+import com.besp.likebesp1.cmnt.service.CmntService;
+import com.besp.likebesp1.common.RsData;
 import com.besp.likebesp1.post.entity.PostDto;
 import com.besp.likebesp1.post.service.PostService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
+import static com.besp.likebesp1.common.LoginUser.LOGIN_USER;
 
 @Controller
 @RequestMapping("/boards")
+@Slf4j
 public class PostController {
 
     private final PostService postService;
     private final BoardService boardService;
-    private static final Logger logger = LoggerFactory.getLogger(PostController.class);
+    private final CmntService cmntService;
 
     @Autowired
-    public PostController(PostService postService, BoardService boardService) {
+    public PostController(PostService postService, BoardService boardService, CmntService cmntService) {
         this.postService = postService;
         this.boardService = boardService;
+        this.cmntService = cmntService;
+    }
+
+    private RsData<String> checkSession(HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        if (session == null) {
+            return RsData.of("F-1", "세션이 만료되었습니다.");
+        }
+
+        long memberId = (long) session.getAttribute(LOGIN_USER.name());
+        return RsData.successOf(String.valueOf(memberId));
+    }
+
+    private boolean canUpdate(long postId, long boardId, String memberId) {
+        PostDto post = postService.getPost(postId, boardId);
+        return post.getMemberId().equals(memberId);
+    }
+
+    private PostDto getVerifiedPost(long boardId, long postId, HttpServletRequest request) {
+        RsData<String> sessionCheckResult = checkSession(request);
+        if (!sessionCheckResult.isSuccess()) {
+            return null;
+        }
+
+        String memberId = sessionCheckResult.getData();
+
+        if (!canUpdate(postId, boardId, memberId)) {
+            return null;
+        }
+
+        return postService.getPost(postId, boardId);
     }
 
     // 게시판에 연결된 게시글 리스트
@@ -44,9 +82,20 @@ public class PostController {
 
     // 게시글 상세 페이지
     @GetMapping("/{boardId}/posts/{postId}")
-    public String getPostDetail(@PathVariable long boardId, @PathVariable long postId, Model model) {
+    public String getPostDetail(@PathVariable long boardId, @PathVariable long postId, Model model, HttpServletRequest request) {
         PostDto post = postService.getPost(postId, boardId);
+        List<CmntDto> cmntList = cmntService.getCmntsByPostId(postId);  // 댓글 목록 가져오기
+
+        // 세션에서 현재 사용자 아이디 가져오기
+        RsData<String> sessionCheckResult = checkSession(request);
+        String currentUserId = sessionCheckResult.getData();
+
+        // 현재 사용자가 게시글 작성자인지 확인
+        boolean isAuthor = post.getMemberId().equals(currentUserId);
+        model.addAttribute("isAuthor", isAuthor);
+
         model.addAttribute("post", post);
+        model.addAttribute("cmntList", cmntList);  // 댓글 목록을 Model에 추가
         return "/post/postDetail";
     }
 
@@ -59,40 +108,47 @@ public class PostController {
     }
 
     @PostMapping("/{boardId}/posts/new")
-    public String writePost(@PathVariable long boardId, @ModelAttribute PostDto postDto) {
+    public String writePost(@PathVariable long boardId, @ModelAttribute PostDto postDto, HttpServletRequest request) {
+        RsData<String> sessionCheckResult = checkSession(request);
+        if (!sessionCheckResult.isSuccess()) {
+            return "redirect:/login";  // 로그인 세션이 없는 경우, 로그인 페이지로 리다이렉트
+        }
+
+        String memberId = sessionCheckResult.getData();
+        postDto.setMemberId(memberId);  // 세션에서 받아온 로그인 아이디를 PostDto에 설정
         long postId = postService.insert(postDto, boardId);  // 삽입된 게시글의 id를 반환받음
         return "redirect:/boards/" + boardId + "/posts/" + postId;  // 게시글 작성 후 해당 게시글의 상세 페이지로 리다이렉트
     }
 
     // 게시글 수정
-    @GetMapping("/{boardId}/posts/{postId}/edit")
-    public String editPostForm(@PathVariable long boardId, @PathVariable long postId, Model model) {
-        PostDto postDto = postService.getPost(boardId, postId);
+    @PatchMapping("/{boardId}/posts/{postId}")
+    @ResponseBody
+    public RsData<String> updatePost(String postTitle, String content, @PathVariable long boardId, @PathVariable long postId, HttpServletRequest request) {
+        PostDto postDto = getVerifiedPost(boardId, postId, request);
+
         if (postDto == null) {
-            logger.error("No post found with boardId: " + boardId + ", postId: " + postId);
-        } else {
-            model.addAttribute("post", postDto);
-            // 모델에 post 객체가 추가되었음을 확인하는 로그
-            logger.info("Post added to model: " + postDto);
+            return RsData.of("F-2", "잘못된 로그인 정보입니다.");
         }
-        return "/post/postEdit";
-    }
 
-
-    @PostMapping("/{boardId}/posts/{postId}/edit")
-    public String editPost(@PathVariable long boardId, @PathVariable long postId, @ModelAttribute PostDto postDto) {
-        postDto.setBoardId(boardId);
-        postDto.setPostId(postId);
+        postDto.setPostTitle(postTitle);
+        postDto.setContent(content);
         postService.updatePost(postDto);
-        return "redirect:/boards/{boardId}/posts/{postId}";
+
+        return RsData.of("S-1", "게시글이 성공적으로 수정되었습니다.");
     }
 
     // 게시글 삭제
     @DeleteMapping("/{boardId}/posts/{postId}")
-    public String deletePost(@PathVariable long boardId, @PathVariable long postId) {
+    @ResponseBody
+    public RsData<String> deletePost(@PathVariable long boardId, @PathVariable long postId, HttpServletRequest request) {
+        PostDto postDto = getVerifiedPost(boardId, postId, request);
+
+        if (postDto == null) {
+            return RsData.of("F-2", "잘못된 로그인 정보입니다.");
+        }
+
         postService.deletePost(postId, boardId);
-        return "redirect:/boards/" + boardId + "/posts";
+
+        return RsData.of("S-1", "게시글이 성공적으로 삭제되었습니다.");
     }
-
-
 }
